requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
pyautogui>=0.9.54
pyyaml>=6.0
pynput>=1.8.0
